Adapted windows build files to new flat folder structure. Fixed 2 warnings for MSVC 2015. Fixed wrong compiler define to decide between windows and other platforms (no idea why this suddenly did not work anymore, maybe due to compiler update).

+++ cc1541.c
@@ -1286,7 +1286,7 @@ generate_uniformat_g64(unsigned char* image, const char *imagepath)
     filepos += write16(track_size, f);
 
     const unsigned int table_size = num_tracks * 4;
-    const unsigned int tracks_offset = filepos + (table_size * 2);
+    const unsigned int tracks_offset = (int)filepos + (table_size * 2);
 
     for (int track = 0; track < num_tracks; ++track) {
         unsigned int track_offset = 0;
